Split behavior impl generic signature tests
