+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -1,3 +1,4 @@
+mod behavior_impl_method_splits;
 mod build_graph_splits;
 mod core_semantics_splits;
 mod declaration_validation_splits;
